// src/components/page/dashboard/types.ts
export type EventData = {
  showAttendingCount: boolean;
  token: string;
  eventDescription: string;
  collectNotComingData: boolean;
  showAttendees: boolean;
  templateId: number;
  collectMaybeData: boolean;
  minimumAgeRequirement: number;
  eventTitle: string;
  collectNote: boolean;
  ageRestricted: boolean;
  attendeeCount?: number;
};